requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.1
Pillow>=11.2.0
python-decouple==3.8
